third_party/blink/web_tests/http/tests/navigation/new-window-redirect-history-expected.txt

This test will open a new window to an empty url and fill its contents with a document.write()
That document will then redirect to another url called redirect-1.html.
That document will redirect to yet another url called redirect-2.html
The back/forward history of the new window should contain both redirect-1.html and redirect-2.html

The test will run automatically with popup blocking disabled, or you can click the button.



============== Back Forward List ==============
curr->  http://127.0.0.1:8000/navigation/new-window-redirect-history.html
===============================================

============== Back Forward List ==============
        http://127.0.0.1:8000/navigation/resources/redirect-1.html
curr->  http://127.0.0.1:8000/navigation/resources/redirect-2.html
===============================================
